drivers: clk: Add default value for HPLL
soc0 SCU010[3:2]:
  00: 2.0GHz (default setting with SCU300, SCU304)
  01: 1.9GHz
  10: 1.8GHz
  11: 1.7GHz (HPLL is not controlled by SCU300, SCU304)

+++ arch/arm/include/asm/arch-aspeed/scu_ast2700.h
@@ -77,14 +77,13 @@
 union ast2700_pll_reg {
 	uint32_t w;
 	struct {
-		unsigned int m : 13;			/* bit[12:0]	*/
-		unsigned int n : 6;			/* bit[18:13]	*/
-		unsigned int p : 4;			/* bit[22:19]	*/
-		unsigned int off : 1;			/* bit[23]	*/
-		unsigned int bypass : 1;		/* bit[24]	*/
-		unsigned int reset : 1;			/* bit[25]	*/
-		unsigned int reserved : 6;		/* bit[31:26]	*/
-
+		uint16_t m : 13;		/* bit[12:0]	*/
+		uint8_t n : 6;			/* bit[18:13]	*/
+		uint8_t p : 4;			/* bit[22:19]	*/
+		uint8_t off : 1;		/* bit[23]	*/
+		uint8_t bypass : 1;		/* bit[24]	*/
+		uint8_t reset : 1;		/* bit[25]	*/
+		uint8_t reserved : 6;		/* bit[31:26]	*/
 	} b;
 };
